// src/app/layout/client/shared/models/prima.model.ts
export class Prima {
  precio: number;
  id: string;
  descripcion: string;
  comisionBroker: number;
  comisionIntermediario: number;
  comisionPuntoVenta: number;
}
export class PrimaFilter {
  TarifaId: string;
  Placa: string;
  Fecha: string;
  Canal: string;
  PuntoVenta: string;
  Cliente: string;
  Departamento: string;
  UsoId: string;
  ClaseId: string;
  MarcaId: string;
  ModeloId: string;
  CantidadAsientos: string;
  TipoPersona: string;
  Moneda: string;
  Carroceria: string;
  CategoriaId: string;
  TipoPapel: number;
  Plan: string;
  BrokerId: string;
  IntermediaId: string;
  SalesPointId: string;
}