// src/app/shared/models/referencechannel/referencechannel.ts
export class ReferenceChannel {
  referenteCode: string;
  referenteCanal: string;
  referenteDesCanal: string;
  referentePuntoVenta: string;
  referenteDesPuntoVenta: string;
  referenteTipoCanal: string;
  referentePlaceholder: string;
  referenteOrigenPublicidad: string;
  referenteAplicaCupon: string;
  referenteStock: string;
  referentePlan: string;
}